refactor: queue message consumer refactoring.
